backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
jinja2==3.1.2
anthropic==0.7.8
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.1
